% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adding_preprocessing_functions.R
\name{add_country}
\alias{add_country}
\title{Adds country column to the dataset}
\usage{
add_country(dataset)
}
\arguments{
\item{dataset}{Dataset to add column to}
}
\value{
The dataset with a new column named "country" added to dataset.
}
\description{
Adds country column to the dataset
}
