% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adding_preprocessing_functions.R
\name{add_country_entity_type}
\alias{add_country_entity_type}
\title{Adds both the entity_type and country columns to the dataset}
\usage{
add_country_entity_type(dataset, type = NA)
}
\arguments{
\item{dataset}{Dataset to add columns to}

\item{type}{Default value to set for the entity.type column}
}
\value{

}
\description{
Adds both the entity_type and country columns to the dataset
}
