% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adding_preprocessing_functions.R
\name{add_entity_type}
\alias{add_entity_type}
\title{Adds the entity_type column to the dataset}
\usage{
add_entity_type(dataset, type)
}
\arguments{
\item{dataset}{Dataset to add column to}

\item{type}{Default entity type to set as value for the column}
}
\value{
Dataset with new column consisting of "type" param added to dataset
**## Not run:**
}
\description{
Adds the entity_type column to the dataset
}
\examples{
add_entity_type(df, "City")
add_entity_type(df, NA)
## End(**Not run**)
}
