% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/cleaning_standardizing_preprocessing_functions.R
\name{clean_country}
\alias{clean_country}
\title{Cleans the dataset's country name to the package's country dictionaries' name}
\usage{
clean_country(dataset, country.dict)
}
\arguments{
\item{dataset}{Dataset to clean the country names for}

\item{country.dict}{Country dictionary to clean the dataset against}
}
\value{
The original dataset with the country names cleaned
}
\description{
Cleans the dataset's country name to the package's country dictionaries' name
}
