% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/cleaning_standardizing_preprocessing_functions.R
\name{clean_country_iso}
\alias{clean_country_iso}
\title{Cleans the dataset's country names and adds iso.}
\usage{
clean_country_iso(dataset, country.dict, iso = 3, clean_enc = T)
}
\arguments{
\item{dataset}{Dataset to clean the country names for}

\item{country.dict}{Country dictionary to clean the dataset against}

\item{iso}{Input either 2 or 3 to select for 2 or 3 letter ISO code. Defaults to ISO3}

\item{clean_enc}{Is the data read in with the correct encoding?
If unknown, set as FALSE. Defaults to TRUE}
}
\value{
The original dataset with the country names cleaned
}
\description{
Cleans the dataset actors' countries using exact matches based
on the package's country dictionary and adds corresponding iso information to
the dataset. For more information on `clean_enc`, see the vignette.
}
