% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/cleaning_standardizing_preprocessing_functions.R
\name{coerce_location_names}
\alias{coerce_location_names}
\title{Coerce location names to handle special characters}
\usage{
coerce_location_names(locations)
}
\arguments{
\item{locations}{Column with names of location (city, country, etc.)}
}
\value{
A vector of locations names with the special character replaced with
closest equivalent
}
\description{
Coerce location names to handle special characters
}
