% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{contextuals}
\alias{contextuals}
\title{Contextuals Database for subnational actors}
\format{contextuals is a dataframe with contextuals information for 10462 unique actors
(rows) and 15 variables (columns)}
\usage{
contextuals
}
\description{
The contextuals database contains important contextual data for the
subnational climate actors found across the different climate databases. See below for
details on what data is included
}
\keyword{datasets}
