% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/cleaning_standardizing_preprocessing_functions.R
\name{.check_and_convert}
\alias{.check_and_convert}
\title{Helper function to help check and convert the encoding of the column specified}
\usage{
.check_and_convert(col)
}
\arguments{
\item{column}{Column to check and convert the encoding for}
}
\value{
column with the encoding (hopefully) converted
}
\description{
Helper function to help check and convert the encoding of the column specified
}
