% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name_phonetic_matching_functions.R
\name{fuzzify_names}
\alias{fuzzify_names}
\title{Cleans name using phonetic matching}
\usage{
fuzzify_names(dataset, key.dict)
}
\arguments{
\item{dataset}{Dataset containing actors by user}

\item{key.dict}{Key dictionary to clean actors' names against}
}
\value{
Cleaned dataset with actors names standardized against the key dictionary.

2 vectors of indices will also be created to store the indices of those names

that needs to be matched.
}
\description{
\code{fuzzify_names} searches for actors' names which are not standardized
and standardizes them according to the key dictionary. Users get to choose which name
is correct based on a selection narrowed down by using phonetic matching. Users can
choose to input custom names should the names not be in the key dictionary.
}
