% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_key_dict.R
\name{load_key_dict}
\alias{load_key_dict}
\title{Download, load, and assign the ClimActor key dictionary to the global environment}
\usage{
load_key_dict()
}
\value{
The function returns \code{invisible(NULL)} and is executed for its
  side effect of assigning \code{key_dict} globally.
}
\description{
This function downloads a specific Parquet file containing a key dictionary
from the 'datadrivenenvirolab/ClimActor' GitHub repository using the
'piggyback' package, loads it into R using 'arrow', and assigns the resulting
data frame to the variable \code{key_dict} in the \strong{global environment}.
}
