% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adding_preprocessing_functions.R
\name{rename_col}
\alias{rename_col}
\title{Renames the "name", "country", and "entity_type" columns to the correct column name}
\usage{
rename_col(dataset, raw_name_col, raw_country_col, raw_type_col)
}
\arguments{
\item{dataset}{Dataset to rename columns for}

\item{raw_name_col}{Current column name for "name"}

\item{raw_country_col}{Current column name for "country"}

\item{raw_type_col}{Current column name for "entity_type"}
}
\value{
The dataset with the appropriately named columns
}
\description{
Renames the "name", "country", and "entity_type" columns to the correct column name
}
