% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/cleaning_standardizing_preprocessing_functions.R
\name{resolve_entity_types}
\alias{resolve_entity_types}
\title{Resolve conflicting entity types}
\usage{
resolve_entity_types(dataset, key.dict, clean_enc = T)
}
\arguments{
\item{dataset}{Dataset containing actors' names}

\item{key.dict}{Key dictionary}

\item{clean_enc}{Is the data read in with the correct encoding?
If unknown, set as FALSE. Defaults to TRUE.}
}
\value{
Returns the dataset with entity types resolved (if applicable)

Also returns a dataset that contains the different conflicts (if applicable)
}
\description{
Resolve entity type conflicts between user's dataset and the key
dictionary. The function checks for entries that have the same name and are from the
same country but have different entity types.between the user's dataset and
the key dictionary. Users can decide whether they would like to accept the key
dictionary's entity types or export a dataframe of the differences for inspection.
}
