% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/cleaning_standardizing_preprocessing_functions.R
\name{standardize_type}
\alias{standardize_type}
\title{Standardizes existing entity types within the dataset}
\usage{
standardize_type(dataset)
}
\arguments{
\item{dataset}{Dataset to standardize and clean the entity type for}
}
\value{
Original dataset with the entity types cleaned
}
\description{
Similar to function \code{fill_type}, guesses the entity type of actors based

based on commonly used words. However, this function applies specifically

to differentiate between different subnational actors (City vs Region)
}
