% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strict_name_match_fun.R
\name{strict_name_match}
\alias{strict_name_match}
\title{Match Names to Climactor_id using and exact match to Key_dict without any extra variables}
\usage{
strict_name_match(x)
}
\arguments{
\item{x}{A data frame containing at least 'name' (the text to be matched) and 'iso' (country code) columns.}
}
\value{
A data frame based on the input \code{x}, but with three new columns added:
  \enumerate{
    \item \code{climactor_id}: The resulting matched identifier.
  }
  The function filters for matches where all lookups results are the same and non-NA \code{climactor_id}.
}
\description{
This function takes a primary data frame to be cleaned (\code{x}) containing at least the 'name' and 'iso' columns
and attempts to match the names to  a 'climactor_id' using the key_dict (\code{key_dict}) primarily its right and wrong columns
It performs several lookups based on original name, modified (lowercase, ASCII)
name, and both 'right' (correct) and 'wrong' (alternative) entries in the dictionary.
It filter for duplicates keeping only the matches where all retrieved \code{climactor_id}s are consistent.
Finally, it joins the retrieved \code{climactor_id} back to the original data

This function requires the \code{key_dict} data frame to be available in the global environment,
and it assumes \code{x} has 'name' and 'iso' columns. Requieres \code{dplyr},
\code{tidyr}, and \code{stringi}.
}
