% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name_phonetic_matching_functions.R
\name{update_country_dict}
\alias{update_country_dict}
\title{Updates the country dictionary}
\usage{
update_country_dict(dataset, country.dict, custom_count)
}
\arguments{
\item{dataset}{Dataset by user}

\item{country.dict}{country dictionary}

\item{custom_count}{Vector of numbers containing which indices were new custom
countries added by the user}
}
\value{
Country dictionary with updated entries
}
\description{
Update the country dictionary with the custom countries given by the
user and the countries that were not in the country dictionary.
}
