% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name_phonetic_matching_functions.R
\name{update_key_dict}
\alias{update_key_dict}
\title{Update key dictionary}
\usage{
update_key_dict(dataset, key.dict, custom_indices)
}
\arguments{
\item{dataset}{Dataset by user}

\item{key.dict}{Key dictionary}

\item{custom_indices}{Vector of numbers containing which indices were new custom names added by the user}
}
\value{
Key dictionary with updated entries
}
\description{
Update the key dictionary with the custom names given by the user and
the names that were not in the key dictionary.
}
