% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/cleaning_standardizing_preprocessing_functions.R
\name{clean_country_iso}
\alias{clean_country_iso}
\title{Cleans the dataset's country names and adds iso.}
\usage{
clean_country_iso(dataset, country.dict, iso = 3, clean_enc = T)
}
\arguments{
\item{dataset}{Dataset to clean the country names for}

\item{country.dict}{Country dictionary to clean the dataset against}

\item{iso}{Input either 2 or 3 to select for 2 or 3 letter ISO code. Defaults to ISO3}

\item{clean_enc}{Is the data read in with the correct encoding?
If unknown, set as FALSE. Defaults to TRUE}
}
\value{
The original dataset with the country names cleaned
}
\description{
Cleans the dataset actors' countries using exact matches based
on the package's country dictionary and adds corresponding iso information to
the dataset. For more information on \code{clean_enc}, see the vignette.
}
\details{
The \code{clean_enc} argument is used for users to denote whether the data is
read in with the correct encoding. Use \code{clean_enc = T} if users are sure of the encoding
that the data is in and is sure data is read in with the right encoding and
\code{clean_enc = F} if users are unsure of data's encoding or if unsure if data is read in
with the correct encoding.

If \code{clean_enc = F}, then the function will check the encoding of the country column in
the user's dataset and tries to fix it if the data was read in with the wrong encoding
(with \code{rvest::repair_encoding()}).
}
