% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/cleaning_standardizing_preprocessing_functions.R
\name{clean_country_iso}
\alias{clean_country_iso}
\title{Cleans the dataset's country names and adds iso.}
\usage{
clean_country_iso(dataset, country.dict, iso = 3, utf = F)
}
\arguments{
\item{dataset}{Dataset to clean the country names for}

\item{country.dict}{Country dictionary to clean the dataset against}

\item{iso}{Input either 2 or 3 to select for 2 or 3 letter ISO code. Defaults to ISO3}

\item{utf}{Is the data in UTF-8 encoding? If unknown, set as FALSE. Defaults to FALSE.}
}
\value{
The original dataset with the country names cleaned
\dontrun{@example clean_country_iso(df, country_dict, iso = 3)}
}
\description{
Cleans the dataset actors' countries based on the package's country
dictionary and adds the corresponding iso to the dataset.
}
