% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name_phonetic_matching_functions.R
\name{clean_name}
\alias{clean_name}
\title{Clean actor names using exact matches}
\usage{
clean_name(dataset, key.dict, clean_enc = T)
}
\arguments{
\item{dataset}{Dataset to clean the actors' names for}

\item{key.dict}{Key dictionary to use to clean the actors' names}

\item{clean_enc}{Is the data read in with the correct encoding?
If unknown, set as FALSE. Defaults to TRUE.}
}
\value{
Returns a dataset with actors names' cleaned using exact string matching.
Also creates a vector of indices of names that require cleaning.
}
\description{
Cleans the datasets' actor names based on the key dictionary using
exact string matching
}
