% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{contextuals}
\alias{contextuals}
\title{Contextuals Database for subnational actors}
\format{
contextuals is a dataframe with contextuals information for 10462 unique actors
(rows) and 15 variables (columns)
#' @format \describe{
    \item{\code{name}}{char Name of the subnational actor}
    \item{\code{iso}}{char 3 letter ISO codes for the country of actor}
    \item{\code{country}}{char Country in which actor resides in}
    \item{\code{entity_type}}{char The entity type of the actor (City, Region, etc.)}
    \item{\code{region}}{char The broader region which the country of the
    subnational actor belongs to}
    \item{\code{area}}{double Total unit area of the actor}
    \item{\code{area_units}}{char Units which the area of the actor are expressed
    in}
    \item{\code{gdp}}{double GDP of the subnational actor}
    \item{\code{initiatives_committed}}{char Climate initiatives to which the actor pledged
    commitments to}
    \item{\code{num_commit}}{int Number of initiatives to which actor pledged commitments
    to}
    \item{\code{lat}}{double Latitude of the actor}
    \item{\code{lng}}{double Longitude of the actor}
    \item{\code{population}}{double Total population of the actor}
    \item{\code{population_year}}{int Year of recorded population}
    \item{\code{state}}{char State in which the actor is situated in}
    }
}
\usage{
contextuals
}
\description{
The contextuals database contains important contextual data for the
subnational climate actors found across the different climate databases. See below for
details on what data is included
}
\keyword{datasets}
