% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/cleaning_standardizing_preprocessing_functions.R
\name{fill_type}
\alias{fill_type}
\title{Fills in the corresponding actor entity type for the dataset based on the
actor name}
\usage{
fill_type(dataset)
}
\arguments{
\item{dataset}{Dataset to fill in the entity type for}
}
\value{
The original dataset with entity types filled for the actors
}
\description{
Guesses the entity type for the actors within the dataset based on commonly
used words for the respective entity types.
}
\details{
The fill type function looks at the actor name and looks for common words
for each entity type (such as "City", "Town" for cities and "Inc.", "Co." for companies)
to guess the entity type for that actor.
}
