% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/cleaning_standardizing_preprocessing_functions.R
\name{remove_extra}
\alias{remove_extra}
\title{Remove extraneous words from actors' names}
\usage{
remove_extra(dataset)
}
\arguments{
\item{dataset}{Dataset containing actors' names}
}
\value{
Dataset with extraneous words removed from actors' names
}
\description{
Removes extraneous words such as "council", "district", etc. from the
names of actors to improve phonetic matching accuracy. See vignette
for the full list of "extraneous" words.
}
\details{
It is recommended that you use this function after the \code{fill_type()} and
\code{resolve_entity_types()} functions, as those functions require the presence of these
"extraneous" words to be fully effective.
}
