% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adding_preprocessing_functions.R
\name{rename_col}
\alias{rename_col}
\title{Renames the columns within the dataset}
\usage{
rename_col(dataset)
}
\arguments{
\item{dataset}{Dataset to rename columns for}
}
\value{
The dataset with the appropriately named columns
**## Not run:**
}
\description{
Checks the dataset for the correct names necessary for further data
processing. If the requisite data names are not in the dataset, then prompts
the user for the columns to be renamed to the requisite names.
}
