% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/cleaning_standardizing_preprocessing_functions.R
\name{standardize_type}
\alias{standardize_type}
\title{Standardizes and cleans existing entity types within the dataset}
\usage{
standardize_type(dataset)
}
\arguments{
\item{dataset}{Dataset to standardize and clean the entity type for}
}
\value{
Original dataset with the entity types cleaned
}
\description{
Standardizes and cleans existing entity types within the dataset
}
